added a notion of bordered and borderless accordion
