+++ packages/react-components/src/accordion/docs/Accordion.stories.mdx
@@ -4,7 +4,6 @@ import { ComponentInfo, Preview, Tagline } from "@stories/components";
 import { Content, Header, InnerContent, InnerHeader } from "@react-components/placeholders";
 import { HelpIcon, InfoIcon, NotificationIcon } from "@react-components/icons";
 import { InnerItem, Item } from "@react-components/collection";
-import { Stack } from "@react-components/layout";
 import { Text } from "@react-components/text";
 
 <Meta
@@ -137,6 +136,29 @@ An accordion items can be rendered dynamically.
 The `expandedKeys` can be retrieved from `useAccordionContext`.
 
 <Preview filePath="/accordion/docs/AccordionContext" />
+
+### Bordered
+
+An accordion can have be contained.
+
+<Preview>
+    <Story name="bordered">
+        <Accordion variant="bordered">
+            <Item key="mars">
+                <Header as="h3">Mars</Header>
+                <Content>Mars is the fourth planet from the Sun and the second-smallest planet in the Solar System.</Content>
+            </Item>
+            <Item key="jupiter">
+                <Header as="h3">Jupiter</Header>
+                <Content>Jupiter is the fifth planet from the Sun and the largest in the Solar System.</Content>
+            </Item>
+            <Item key="venus">
+                <Header as="h3">Venus</Header>
+                <Content>Venus is the second planet from the Sun. It is named after the Roman goddess of love and beauty.</Content>
+            </Item>
+        </Accordion>
+    </Story>
+</Preview>
 
 ### Disabled
