+++ packages/react-components/src/accordion/src/Accordion.css
@@ -5,10 +5,22 @@
     width: 100%;
 }
 
+/* BORDERED */
+.o-ui-accordion-bordered {
+    box-shadow: var(--o-ui-alias-elevation-skim);
+    border-radius: var(--o-ui-global-scale-bravo);
+    background-color: var(--o-ui-alias-background-1);
+}
+
 /* HEADER */
 .o-ui-accordion-header {
     margin-bottom: 0;
-    margin-top: var(--o-ui-global-scale-delta);
+}
+
+/* HEADER BORDERLESS */
+.o-ui-accordion-borderless .o-ui-accordion-header .o-ui-accordion-trigger {
+    padding-top: var(--o-ui-global-scale-bravo);
+    padding-bottom: var(--o-ui-global-scale-bravo);
 }
 
 .o-ui-accordion-header:first-child {
@@ -21,16 +33,29 @@
     display: flex;
     align-items: center;
     text-align: left;
-    background-color: var(--o-ui-alias-background-1);
     cursor: pointer;
+    background-color: transparent;
     border: 0;
-    transition: background-color .3s ease-in, border-bottom-left-radius 5s linear, border-bottom-right-radius  5s linear;
+    transition: background-color .3s ease-in, border-bottom-left-radius 5s linear, border-bottom-right-radius 5s linear;
     width: 100%;
     color: var(--o-ui-alias-text-2);
     line-height: inherit;
     font-weight: inherit;
     position: relative;
     padding: 0 var(--o-ui-global-scale-bravo);
+}
+
+/* TRIGGER | BORDERED */
+.o-ui-accordion-bordered .o-ui-accordion-trigger {
+    background-color: var(--o-ui-alias-background-1);
+    padding: calc(var(--o-ui-global-scale-delta) - 1px) var(--o-ui-global-scale-echo) var(--o-ui-global-scale-delta) var(--o-ui-global-scale-foxtrot);
+    border-top: 1px solid var(--o-ui-alias-border-2);
+}
+
+/* TRIGGER | BORDERED | HOVER */
+.o-ui-accordion .o-ui-accordion-trigger:not([disabled]):hover,
+.o-ui-accordion .o-ui-accordion-trigger-hover:not([disabled]) {
+    background-color: var(--o-ui-alias-background-accordion-hover);
 }
 
 /* TRIGGER | CONTENT */
@@ -46,7 +71,11 @@
     line-height: 1;
 }
 
-/* TRIGGER | CONTENT | TITLE | ICON */
+/* TRIGGER | BORDERED | HAS ICON */
+.o-ui-accordion-bordered .o-ui-accordion-trigger.o-ui-accordion-trigger-has-icon {
+    padding-left: var(--o-ui-global-scale-delta);
+}
+
 .o-ui-accordion-icon + .o-ui-accordion-title {
     margin-left: var(--o-ui-global-scale-alpha);
 }
@@ -55,8 +84,18 @@
 .o-ui-accordion-panel {
     color: var(--o-ui-alias-text-1);
     background-color: var(--o-ui-alias-background-1);
+}
+
+/* PANEL | BORDERLESS */
+.o-ui-accordion-borderless .o-ui-accordion-panel {
     padding-left: var(--o-ui-global-scale-bravo);
     padding-right: var(--o-ui-global-scale-bravo);
+}
+
+/* PANEL | BORDERED */
+.o-ui-accordion-bordered .o-ui-accordion-panel {
+    padding-left: var(--o-ui-global-scale-foxtrot);
+    padding-right: calc(var(--o-ui-global-scale-foxtrot) + var(--o-ui-global-scale-foxtrot));
 }
 
 /* STATES */
@@ -89,10 +128,22 @@
     border: var(--o-ui-focus-ring-thickness-md) transparent solid;
     border-radius: var(--o-ui-shape-rounded);
     pointer-events: none;
+}
+
+/* FOCUS | BORDERLESS */
+.o-ui-accordion-borderless .o-ui-accordion-trigger::before {
     left: 0;
     right: 0;
     bottom: 0;
     top: 0;
+}
+
+/* FOCUS | BORDERED */
+.o-ui-accordion-bordered .o-ui-accordion-trigger::before {
+    left: var(--o-ui-global-scale-delta);
+    right: var(--o-ui-global-scale-delta);
+    bottom: var(--o-ui-global-scale-bravo);
+    top: var(--o-ui-global-scale-bravo);
 }
 
 .o-ui-accordion-trigger:focus-visible,
@@ -104,4 +155,25 @@
 .o-ui-accordion-trigger:focus-visible::before,
 .o-ui-accordion-trigger-focus::before {
     border-color: var(--o-ui-alias-border-primary-1);
+}
+
+/* BORDERED | FIRST ITEM | BORDER RADIUS */
+.o-ui-accordion-bordered .o-ui-accordion-header:first-of-type,
+.o-ui-accordion-bordered .o-ui-accordion-header:first-of-type .o-ui-accordion-trigger {
+    border-top-left-radius: var(--o-ui-global-scale-bravo);
+    border-top-right-radius: var(--o-ui-global-scale-bravo);
+}
+
+/* BORDERED | LAST ITEM | BORDER RADIUS */
+.o-ui-accordion-bordered .o-ui-accordion-header:last-of-type,
+.o-ui-accordion-bordered .o-ui-accordion-header:last-of-type .o-ui-accordion-trigger {
+    border-bottom-left-radius: var(--o-ui-global-scale-bravo);
+    border-bottom-right-radius: var(--o-ui-global-scale-bravo);
+}
+
+/* BORDERED | LAST PANEL | BORDER RADIUS */
+.o-ui-accordion-bordered .o-ui-accordion-panel:last-of-type,
+.o-ui-accordion-bordered .o-ui-accordion-panel:last-of-type .o-ui-accordion-trigger {
+    border-bottom-left-radius: var(--o-ui-global-scale-bravo);
+    border-bottom-right-radius: var(--o-ui-global-scale-bravo);
 }
