+++ packages/react-components/src/accordion/src/Accordion.tsx
@@ -43,6 +43,10 @@ export interface InnerAccordionProps extends DomProps {
      */
     expansionMode: "single" | "multiple";
     /**
+     * The type of expansion that is allowed.
+     */
+    variant: "bordered" | "borderless";
+    /**
      * Whether or not the first focusable accordion item should autoFocus on render.
      */
     autoFocus?: boolean | number;
@@ -67,6 +71,7 @@ export function InnerAccordion({
     onExpansionChange,
     expansionMode = "single",
     autoFocus,
+    variant = "borderless",
     as = "div",
     children,
     forwardedRef,
@@ -119,7 +124,7 @@ export function InnerAccordion({
             {...mergeProps(
                 rest,
                 {
-                    className: "o-ui-accordion",
+                    className: `o-ui-accordion o-ui-accordion-${variant}`,
                     as,
                     ref: containerRef
                 },
