+++ packages/react-components/src/button/tests/chromatic/Button.chroma.jsx
@@ -25,6 +25,3 @@ stories()
             <Button style={{ backgroundColor: "red" }}>Button</Button>
         </Inline>
     );
-
-
-
